refactor: split MCP transport runtime layers
- split MCP client layout into manager and transport modules
- add core MCP runtime service for sampling orchestration
- route streaming MCP actions through runtime input interface
- slim event loop MCP helpers by using runtime utility functions
- update architecture docs for new MCP module structure
