+++ src/core/app/actions/streaming.rs
@@ -4,22 +4,17 @@ use std::time::Instant;
 use super::{App, AppAction, AppActionContext, AppCommand};
 use crate::api::{ChatMessage, ChatToolCall, ChatToolCallFunction};
 use crate::core::app::session::{
-    McpPromptRequest, McpSamplingRequest, PendingToolCall, ToolCallRequest, ToolFailureKind,
-    ToolPayloadHistoryEntry, ToolResultRecord, ToolResultStatus,
+    McpPromptRequest, PendingToolCall, ToolCallRequest, ToolFailureKind, ToolPayloadHistoryEntry,
+    ToolResultRecord, ToolResultStatus,
 };
 use crate::core::app::ui_state::ToolPromptRequest;
 use crate::core::chat_stream::StreamParams;
 use crate::core::config::data::McpToolPayloadRetention;
-use crate::core::mcp_sampling::{
-    build_sampling_messages, serialize_sampling_params, summarize_sampling_request,
-};
+use crate::core::mcp_runtime::{handle_runtime_input, McpRuntimeInput};
 use crate::core::message::{AppMessageKind, Message, ROLE_ASSISTANT, ROLE_USER};
 use crate::mcp::permissions::ToolPermissionDecision;
 use crate::mcp::{MCP_INSTANT_RECALL_TOOL, MCP_SESSION_MEMORY_SERVER_ID};
-use rust_mcp_schema::schema_utils::ServerJsonrpcRequest;
-use rust_mcp_schema::{
-    ClientCapabilities, ClientSampling, ContentBlock, PromptMessage, Role, RpcError,
-};
+use rust_mcp_schema::{ContentBlock, PromptMessage, Role};
 use serde_json::{Map, Value};
 use tracing::debug;
 
@@ -103,9 +98,11 @@ pub(super) fn handle_streaming_action(
             handle_mcp_prompt_completed(app, request, result, ctx)
         }
         AppAction::McpServerRequestReceived { request } => {
-            handle_mcp_server_request(app, *request, ctx)
+            handle_runtime_input(app, McpRuntimeInput::ServerRequest(*request), ctx)
         }
-        AppAction::McpSamplingFinished => handle_mcp_sampling_finished(app, ctx),
+        AppAction::McpSamplingFinished => {
+            handle_runtime_input(app, McpRuntimeInput::SamplingFinished, ctx)
+        }
         AppAction::StreamErrored { message, stream_id } => {
             if !app.is_current_stream(stream_id) {
                 return None;
@@ -398,7 +395,11 @@ fn handle_tool_permission_decision(
 
     if prompt_tool.as_deref() == Some(crate::mcp::MCP_SAMPLING_TOOL) {
         let request = app.session.active_sampling_request.take()?;
-        return handle_sampling_permission_decision(app, request, decision, ctx);
+        return handle_runtime_input(
+            app,
+            McpRuntimeInput::SamplingPermission { request, decision },
+            ctx,
+        );
     }
 
     if let Some(request) = app.session.active_tool_request.take() {
@@ -452,155 +453,11 @@ fn handle_tool_permission_decision(
     }
 
     let request = app.session.active_sampling_request.take()?;
-    handle_sampling_permission_decision(app, request, decision, ctx)
-}
-
-fn handle_mcp_server_request(
-    app: &mut App,
-    request: crate::mcp::events::McpServerRequest,
-    ctx: AppActionContext,
-) -> Option<AppCommand> {
-    debug!(
-        server_id = %request.server_id,
-        request_id = ?request.request.request_id(),
-        method = %request.request.method(),
-        "Received MCP server request"
-    );
-    if app.session.mcp_disabled {
-        debug!(
-            server_id = %request.server_id,
-            request_id = ?request.request.request_id(),
-            "MCP server request ignored (MCP disabled)"
-        );
-        return Some(AppCommand::SendMcpServerError {
-            server_id: request.server_id,
-            request_id: request.request.request_id().clone(),
-            error: RpcError::internal_error().with_message("MCP is disabled."),
-        });
-    }
-
-    let capabilities = ClientCapabilities {
-        sampling: Some(ClientSampling::default()),
-        ..ClientCapabilities::default()
-    };
-    if let Err(err) = capabilities.can_handle_request(&request.request) {
-        debug!(
-            server_id = %request.server_id,
-            request_id = ?request.request.request_id(),
-            error_code = err.code,
-            "MCP server request rejected (unsupported)"
-        );
-        return Some(AppCommand::SendMcpServerError {
-            server_id: request.server_id,
-            request_id: request.request.request_id().clone(),
-            error: err,
-        });
-    }
-
-    let request_id = request.request.request_id().clone();
-    let ServerJsonrpcRequest::CreateMessageRequest(create_request) = request.request else {
-        debug!(
-            server_id = %request.server_id,
-            request_id = ?request_id,
-            "MCP server request rejected (unexpected method)"
-        );
-        return Some(AppCommand::SendMcpServerError {
-            server_id: request.server_id,
-            request_id,
-            error: RpcError::method_not_found()
-                .with_message("Unsupported MCP request from server."),
-        });
-    };
-
-    let messages = match build_sampling_messages(&create_request) {
-        Ok(messages) => messages,
-        Err(err) => {
-            debug!(
-                server_id = %request.server_id,
-                request_id = ?create_request.id,
-                error = %err,
-                "MCP sampling request rejected (invalid params)"
-            );
-            return Some(AppCommand::SendMcpServerError {
-                server_id: request.server_id,
-                request_id: create_request.id.clone(),
-                error: RpcError::invalid_params().with_message(&err),
-            });
-        }
-    };
-
-    app.session
-        .pending_sampling_queue
-        .push_back(McpSamplingRequest {
-            server_id: request.server_id,
-            request: create_request,
-            messages,
-        });
-
-    debug!(
-        pending = app.session.pending_sampling_queue.len(),
-        active = app.session.active_sampling_request.is_some(),
-        "Queued MCP sampling request"
-    );
-
-    advance_sampling_queue(app, ctx)
-}
-
-fn handle_sampling_permission_decision(
-    app: &mut App,
-    request: McpSamplingRequest,
-    decision: ToolPermissionDecision,
-    ctx: AppActionContext,
-) -> Option<AppCommand> {
-    debug!(
-        server_id = %request.server_id,
-        request_id = ?request.request.id,
-        decision = ?decision,
-        "Sampling permission decision"
-    );
-    app.ui.cancel_tool_prompt();
-    app.clear_status();
-
-    match decision {
-        ToolPermissionDecision::AllowOnce => {}
-        ToolPermissionDecision::AllowSession | ToolPermissionDecision::Block => app
-            .mcp_permissions
-            .record(&request.server_id, crate::mcp::MCP_SAMPLING_TOOL, decision),
-        ToolPermissionDecision::DenyOnce => {}
-    }
-
-    if matches!(
-        decision,
-        ToolPermissionDecision::DenyOnce | ToolPermissionDecision::Block
-    ) {
-        let message = match decision {
-            ToolPermissionDecision::Block => "Sampling blocked by user.",
-            _ => "Sampling denied by user.",
-        };
-        return Some(AppCommand::SendMcpServerError {
-            server_id: request.server_id,
-            request_id: request.request.id.clone(),
-            error: RpcError {
-                code: -1,
-                message: message.to_string(),
-                data: None,
-            },
-        });
-    }
-
-    app.session.active_sampling_request = Some(request.clone());
-    set_status_for_sampling_run(app, &request, ctx);
-    Some(AppCommand::RunMcpSampling(Box::new(request)))
-}
-
-fn handle_mcp_sampling_finished(app: &mut App, ctx: AppActionContext) -> Option<AppCommand> {
-    app.session.active_sampling_request = None;
-    if let Some(request) = app.session.active_tool_request.clone() {
-        set_status_for_tool_run(app, &request, ctx);
-    } else {
-        app.end_mcp_operation_if_active();
-    }
-    advance_sampling_queue(app, ctx)
+    handle_runtime_input(
+        app,
+        McpRuntimeInput::SamplingPermission { request, decision },
+        ctx,
+    )
 }
 
 fn handle_tool_call_completed(
@@ -1062,87 +919,6 @@ fn advance_tool_queue(app: &mut App, ctx: AppActionContext) -> Option<AppCommand
         args_summary,
         raw_arguments,
         batch_index,
-    });
-    None
-}
-
-fn advance_sampling_queue(app: &mut App, ctx: AppActionContext) -> Option<AppCommand> {
-    debug!(
-        pending = app.session.pending_sampling_queue.len(),
-        active = app.session.active_sampling_request.is_some(),
-        "Advance MCP sampling queue"
-    );
-    if app.session.active_sampling_request.is_some() {
-        return None;
-    }
-
-    if app.ui.tool_prompt().is_some() {
-        return None;
-    }
-
-    let request = app.session.pending_sampling_queue.pop_front()?;
-    debug!(
-        server_id = %request.server_id,
-        request_id = ?request.request.id,
-        pending = app.session.pending_sampling_queue.len(),
-        "Dequeued MCP sampling request"
-    );
-
-    if is_mcp_yolo_enabled(app, &request.server_id) {
-        app.session.active_sampling_request = Some(request.clone());
-        set_status_for_sampling_run(app, &request, ctx);
-        return Some(AppCommand::RunMcpSampling(Box::new(request)));
-    }
-
-    if let Some(decision) = app
-        .mcp_permissions
-        .decision_for(&request.server_id, crate::mcp::MCP_SAMPLING_TOOL)
-    {
-        if decision == ToolPermissionDecision::Block {
-            return Some(AppCommand::SendMcpServerError {
-                server_id: request.server_id,
-                request_id: request.request.id.clone(),
-                error: RpcError {
-                    code: -1,
-                    message: "Sampling blocked by user.".to_string(),
-                    data: None,
-                },
-            });
-        }
-
-        if decision == ToolPermissionDecision::AllowSession {
-            debug!(
-                server_id = %request.server_id,
-                request_id = ?request.request.id,
-                "Auto-allowing sampling for session"
-            );
-            app.session.active_sampling_request = Some(request.clone());
-            set_status_for_sampling_run(app, &request, ctx);
-            return Some(AppCommand::RunMcpSampling(Box::new(request)));
-        }
-    }
-
-    let server_name = resolve_server_label(app, &request.server_id);
-    let args_summary = summarize_sampling_request(&request.request);
-    let raw_arguments = serialize_sampling_params(&request.request);
-    debug!(
-        server_id = %request.server_id,
-        request_id = ?request.request.id,
-        "Prompting user for sampling permission"
-    );
-    app.session.active_sampling_request = Some(request.clone());
-    let display_name = format!(
-        "Allow {} to generate text with {}?",
-        server_name, app.session.model
-    );
-    app.ui.start_tool_prompt(ToolPromptRequest {
-        server_id: request.server_id,
-        server_name,
-        tool_name: crate::mcp::MCP_SAMPLING_TOOL.to_string(),
-        display_name: Some(display_name),
-        args_summary,
-        raw_arguments,
-        batch_index: 0,
     });
     None
 }
@@ -1626,6 +1402,13 @@ fn build_tool_result_summary(
     summary
 }
 
+fn is_mcp_yolo_enabled(app: &App, server_id: &str) -> bool {
+    app.mcp
+        .server(server_id)
+        .map(|server| server.config.is_yolo())
+        .unwrap_or(false)
+}
+
 fn set_status_for_tool_run(app: &mut App, _request: &ToolCallRequest, ctx: AppActionContext) {
     let input_area_height = app.input_area_height(ctx.term_width);
     let _token = app.begin_mcp_operation();
@@ -1633,26 +1416,6 @@ fn set_status_for_tool_run(app: &mut App, _request: &ToolCallRequest, ctx: AppAc
     let available_height =
         conversation.calculate_available_height(ctx.term_height, input_area_height);
     conversation.update_scroll_position(available_height, ctx.term_width);
-}
-
-fn set_status_for_sampling_run(
-    app: &mut App,
-    _request: &McpSamplingRequest,
-    ctx: AppActionContext,
-) {
-    let input_area_height = app.input_area_height(ctx.term_width);
-    let _token = app.begin_mcp_operation();
-    let mut conversation = app.conversation();
-    let available_height =
-        conversation.calculate_available_height(ctx.term_height, input_area_height);
-    conversation.update_scroll_position(available_height, ctx.term_width);
-}
-
-fn is_mcp_yolo_enabled(app: &App, server_id: &str) -> bool {
-    app.mcp
-        .server(server_id)
-        .map(|server| server.config.is_yolo())
-        .unwrap_or(false)
 }
 
 fn resolve_server_label(app: &App, server_id: &str) -> String {
